sync: hosted cloud (#1385)
Submit a pull request for this project.

<!-- If you have an Issue that related to this Pull Request, you can
copy this Issue's description -->

# Why? 
<!-- 
> Related to which issue?
> Why we need this pull request?
> What is the user story for this pull request? 
-->


# What?
<!-- 
> Can you describe this feature in detail?
> Who can benefit from it? 
-->


# How?
<!-- 
> Do you have a simple description of how this pull request is
implemented?
-->

---------

+++ backend-server/application/src/main/java/com/apitable/automation/entity/AutomationTriggerEntity.java
@@ -18,15 +18,15 @@
 
 package com.apitable.automation.entity;
 
-import java.io.Serializable;
-import java.time.LocalDateTime;
-
 import com.baomidou.mybatisplus.annotation.FieldFill;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableLogic;
 import com.baomidou.mybatisplus.annotation.TableName;
+import java.io.Serializable;
+import java.math.BigInteger;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -56,7 +56,7 @@ public class AutomationTriggerEntity implements Serializable {
      * Primary Key
      */
     @TableId(value = "id", type = IdType.ASSIGN_ID)
-    private Long id;
+    private BigInteger id;
 
     /**
      * Robot ID (link#xxxx_automation_robot#robot_id)
